Chapter 2 Placements
Important: the only non-placement is to the title of "The Chain Rule". It was "The Chain Rule", but I changed it to "Chain Rule."
In the text, there were many references to "the Chain Rule." However, some PreTeXt edits made these into hyperlinks and the text became "the The Chain Rule." I think this should fix should work well, at it mimics other titles such as "Product Rule" vs. "The Product Rule."